+++ src/methods/mulmo_media_source.ts
@@ -1,6 +1,21 @@
 import fs from "fs";
 import { MulmoMediaSource, MulmoMediaMermaidSource, MulmoStudioContext } from "../types/index.js";
-import { getFullPath, resolveAssetPath } from "../utils/file.js";
+import { getFullPath, getReferenceImagePath, resolveAssetPath } from "../utils/file.js";
+import { getExtention } from "../utils/utils.js";
+
+const downLoadReferenceImage = async (context: MulmoStudioContext, key: string, url: string) => {
+  const response = await fetch(url);
+  if (!response.ok) {
+    throw new Error(`Failed to download image: ${url}`);
+  }
+  const buffer = Buffer.from(await response.arrayBuffer());
+
+  // Detect file extension from Content-Type header or URL
+  const extension = getExtention(response.headers.get("content-type"), url);
+  const imagePath = getReferenceImagePath(context, key, extension);
+  await fs.promises.writeFile(imagePath, buffer);
+  return imagePath;
+};
 
 export const MulmoMediaSourceMethods = {
   async getText(mediaSource: MulmoMediaMermaidSource, context: MulmoStudioContext) {
@@ -29,5 +44,15 @@ export const MulmoMediaSourceMethods = {
       return mediaSource.url;
     }
     return null;
+  },
+  // if url then download image and save it to file. both case return local image path.
+  async imageReference(mediaSource: MulmoMediaSource, context: MulmoStudioContext, key: string) {
+    if (mediaSource.kind === "path") {
+      return resolveAssetPath(context, mediaSource.path);
+    } else if (mediaSource.kind === "url") {
+      return await downLoadReferenceImage(context, key, mediaSource.url);
+    }
+    // TODO base64
+    throw new Error(`imageReference media unknown error`); // TODO cause
   },
 };
